revert bot changes.
