+++ src/Devices/Bot.ts
@@ -1,10 +1,7 @@
-import {
-  CharacteristicEventTypes,
-  CharacteristicGetCallback,
-  PlatformAccessory,
-  Service,
-} from 'homebridge';
+import { Service, PlatformAccessory, CharacteristicEventTypes, CharacteristicSetCallback } from 'homebridge';
 import { SwitchBotPlatform } from '../platform';
+import { interval, Subject } from 'rxjs';
+import { debounceTime, skipWhile, tap } from 'rxjs/operators';
 import { DeviceURL } from '../settings';
 import { device } from '../configTypes';
 
@@ -14,17 +11,31 @@ import { device } from '../configTypes';
  * Each accessory may expose multiple services of different service types.
  */
 export class Bot {
-  private service!: Service;
+  private service: Service;
 
   On!: boolean;
   OutletInUse!: boolean;
   deviceStatus!: any;
+
+  botUpdateInProgress!: boolean;
+  doBotUpdate!: any;
 
   constructor(
     private readonly platform: SwitchBotPlatform,
     private accessory: PlatformAccessory,
     public device: device,
   ) {
+    // default placeholders
+    this.On = false;
+    this.OutletInUse = true;
+
+    // this is subject we use to track when we need to POST changes to the SwitchBot API
+    this.doBotUpdate = new Subject();
+    this.botUpdateInProgress = false;
+
+    // Retrieve initial values and updateHomekit
+    this.parseStatus();
+
     // set accessory information
     this.accessory
       .getService(this.platform.Service.AccessoryInformation)!
@@ -32,7 +43,7 @@ export class Bot {
       .setCharacteristic(this.platform.Characteristic.Model, 'SWITCHBOT-BOT-S1')
       .setCharacteristic(this.platform.Characteristic.SerialNumber, this.device.deviceId);
 
-    // get the Television service if it exists, otherwise create a new Television service
+    // get the LightBulb service if it exists, otherwise create a new LightBulb service
     // you can create multiple services for each accessory
     (this.service =
       this.accessory.getService(this.platform.Service.Outlet) ||
@@ -50,17 +61,80 @@ export class Bot {
       `${this.device.deviceName} ${this.device.deviceType}`,
     );
 
-    // handle on / off events using the Active characteristic
+    // each service must implement at-minimum the "required characteristics" for the given service type
+    // see https://developers.homebridge.io/#/service/Outlet
+
     this.service
       .getCharacteristic(this.platform.Characteristic.On)
-      .on(CharacteristicEventTypes.SET, (value: any, callback: CharacteristicGetCallback) => {
-        this.platform.log.debug('Bot %s Set On: %s', this.accessory.displayName, value);
-        this.platform.log.warn(value);
-        this.pushChanges();
-        this.On = value;
-        this.service.updateCharacteristic(this.platform.Characteristic.On, this.On);
-        callback(null);
+      .on(CharacteristicEventTypes.SET, this.handleOnSet.bind(this));
+
+    // Retrieve initial values and updateHomekit
+    this.updateHomeKitCharacteristics();
+
+    // Start an update interval
+    interval(this.platform.config.options!.refreshRate! * 1000)
+      .pipe(skipWhile(() => this.botUpdateInProgress))
+      .subscribe(() => {
+        this.refreshStatus();
       });
+
+    // Watch for Bot change events
+    // We put in a debounce of 100ms so we don't make duplicate calls
+    this.doBotUpdate
+      .pipe(
+        tap(() => {
+          this.botUpdateInProgress = true;
+        }),
+        debounceTime(100),
+      )
+      .subscribe(async () => {
+        try {
+          await this.pushChanges();
+        } catch (e) {
+          this.platform.log.error(JSON.stringify(e.message));
+          this.platform.log.debug('Bot %s -', this.accessory.displayName, JSON.stringify(e));
+        }
+        this.botUpdateInProgress = false;
+      });
+  }
+
+  /**
+   * Parse the device status from the SwitchBot api
+   */
+  parseStatus() {
+    this.OutletInUse = true;
+    if (this.platform.config.options?.bot?.device_press?.includes(this.device.deviceId)) {
+      this.On = false;
+    }
+    this.platform.log.debug('Bot %s OutletInUse: %s On: %s', this.accessory.displayName, this.OutletInUse, this.On);
+  }
+
+  /**
+   * Asks the SwitchBot API for the latest device information
+   */
+  async refreshStatus() {
+    try {
+      // this.platform.log.error('Bot - Reading', `${DeviceURL}/${this.device.deviceID}/devices`);
+      const deviceStatus: any = {
+        statusCode: 100,
+        body: {
+          deviceId: this.device.deviceId,
+          deviceType: this.device.deviceType,
+          hubDeviceId: this.device.hubDeviceId,
+          power: 'on',
+        },
+        message: 'success',
+      };
+      this.deviceStatus = deviceStatus;
+      this.parseStatus();
+      this.updateHomeKitCharacteristics();
+    } catch (e) {
+      this.platform.log.error(
+        `Bot - Failed to update status of ${this.device.deviceName}`,
+        JSON.stringify(e.message),
+        this.platform.log.debug('Bot %s -', this.accessory.displayName, JSON.stringify(e)),
+      );
+    }
   }
 
   /**
@@ -107,5 +181,25 @@ export class Bot {
     // Make the API request
     const push = await this.platform.axios.post(`${DeviceURL}/${this.device.deviceId}/commands`, payload);
     this.platform.log.debug('Bot %s Changes pushed -', this.accessory.displayName, push.data);
+    this.refreshStatus();
+  }
+
+  /**
+   * Updates the status for each of the HomeKit Characteristics
+   */
+  updateHomeKitCharacteristics() {
+    this.service.updateCharacteristic(this.platform.Characteristic.On, this.On);
+    this.service.updateCharacteristic(this.platform.Characteristic.OutletInUse, this.OutletInUse);
+  }
+
+  /**
+   * Handle requests to set the "On" characteristic
+   */
+  handleOnSet(value: any, callback: CharacteristicSetCallback) {
+    this.platform.log.debug('Bot %s -', this.accessory.displayName, `Set On: ${value}`);
+    this.doBotUpdate.next();
+    this.On = value;
+    this.service.updateCharacteristic(this.platform.Characteristic.On, this.On);
+    callback(null);
   }
 }
